requirements.txt
aiohttp~=3.9.3
aiosqlite~=0.20.0
discord.py
python-dotenv~=1.0.1